Prefer requestAnimationFrame over setInterval
`requestAnimationFrame` is preferred over `setInterval` for drawing
frames as it allows the browser itself to make decisions on when to draw
the layout. Unfortunately, we are no longer capped at 30fps which will
result in higher CPU usage, but I feel the tradeoff is acceptable here
as the browser will throttle itself if needed.

On the plus side however, input latency will be lower overall and
frametimes will be more consistent.