[CI] Don't pad the current week number
The default formatting pads the current week as 00..53. This works for every week other than the current two weeks in the year as those are formatted as 08 and 09. The shell however interprets those as integers in base 8 (octal) because of the initial 0 (blame C for this one), where 8 and 9 are not valid digits in base 8.

This commit removes the padding with 0 from the formatting.