Implement Support for Abbreviations
This implements abbreviations. For the longest time we thought it's not
possible to implement them in the web, but it turns out that it can
work.

This implements support for the key value pair components so far.
Abbreviations are still missing for the Title component. This will need
some additional support from livesplit-core.