Store the splits as a byte array
From now on we never serialize the splits to a string and instead
directly slice the WebAssembly memory and directly export, upload, and
save the splits from there.

Additionally this does another optimization where livesplit-core now
provides a way to directly query the length of the returned buffer
instead of having to manually search of the nul-terminator.