Inline the main JS bundle into the HTML
By inlining a tiny amount of JS into the HTML, we can request the WASM
and other JS bundles much sooner, reducing the latency a bit. This also
inlines the most important styling, so that the loading screen is always
properly styled. We unfortunately have to use `syncWebAssembly` here as
it allows requesting the WASM file sooner than the other setting. It's
not a whole lot sooner, so once they remove the setting, we can switch
back without taking too much of a hit.