Use smaller bundles again
Turns out that compiling a large bundle is really slow and apparently
also not parallelized. I thought it might make sense to not split the
bundle to not create any unnecessary requests. But JavaScript is
apparently really slow to compile, so any sort of parallelization we can
get definitely outweighs the slight cost to the additional requests.