Fix Hotkeys not being set properly
When changing the hotkeys in the settings, it doesn't properly set the
hotkeys anymore. This is because of a refactor that we merged in
`livesplit-core` that unregisters the hotkeys while the `HotkeySystem`
is deactivated. The hotkey settings deactivate the hotkeys so you
don't accidentally split / reset / ... the timer in the background. The
bug was that it was completely impossible to set hotkeys while
they are deactivated.