Implement Hotkey Modifiers (#740)
This adds support for hotkey modifiers, such as Ctrl, Alt, Shift and the
Meta key. On top of that I ran `npm update` to truly update all the
packages to their latest semver compatible versions, which `dependabot`
doesn't seem to do for us. This also includes other updates in
`livesplit-core` that were done since the last update, such as the
precise formatting and parsing of the time stamps.