Resolve Key Codes with `livesplit-core` (#748)
This updates `livesplit-core` such that we can resolve the key codes
with its built-in resolving code. Also since `livesplit-core` introduced
a `max-opt` profile, we make use of that now and on top of that also
introduce `virtual-function-elimination` for the nightly / CI builds.
This is an additional LLVM pass that tries to eliminate as many function
pointers as possible by analyzing the whole program. Locally this
reduced the size of the function pointer table from around 500 to around
400 entries. There's also a flag for using WebAssembly SIMD
instructions, as that works perfectly fine on all browsers except
Safari. Of course we won't use it for deploying until Safari supports
those instructions, but it may be useful for some local testing.

Also with this `livesplit-core` update `SpeedRunIGT` splits can now be
parsed.