Implement Screen Wake Lock Support
This attempts to use the Screen Wake Lock API to prevent the screen from
turning off. This is available in iOS 16.4 now, however at least when
adding the app to the home screen as a PWA, the functionality seems to
be broken. Regardless, that's not something we can fix. We can instead
simply always try to activate it and hopefully Apple fixes the bug soon.

https://bugs.webkit.org/show_bug.cgi?id=254545