Query local fonts to show a font list (#860)
This uses the [Local Font Access
API](https://developer.mozilla.org/en-US/docs/Web/API/Local_Font_Access_API)
to query the list of fonts installed on the user's system. If we
successfully query it we show a combobox with the list of fonts.
Otherwise it's still a textbox. We also get information about the styles
that are available for each font family. Unfortunately this comes as a
string rather than individual fields, so we have to parse it. We do this
by looking for specific style names in there. Since this is mostly
heuristic based, instead of only showing the styles that we found, we
still show all of them, but grey out those that we didn't find, but you
can still select them.