Allow toggling speedrun.com, splits.io integration (#886)
This adds settings for turning off the integration with speedrun.com and
splits.io. This is useful for people who don't want to use these
features and want to avoid the extra network requests. While splits.io
doesn't do any automatic network requests in the background, it still
adds buttons to the UI that someone may not want to see. It's also more
consistent to have a toggle for both.