Enable Multivalue again
With LLVM 19 they explicitly require using the `experimental-mv` ABI in
addition to the `multivalue` LLVM feature (the latter is enabled by
default now). Enabling the ABI is non-trivial and requires a full on
custom Rust target. This adds `wasm32-multivalue.json` which is
identical to `wasm32-unknown-unknown` except it sets the
`experimental-mv` ABI. The target JSON format is very experimental by
itself, so we may need to keep this up to date from time to time.