Turn On WebAssembly Tail Calls
With Safari 18.2 having shipped today, all browsers now support
WebAssembly Tail Calls. This allows the compiler to call a function at
the end of the current function by reusing the stack frame. This should
be a tiny bit faster and allow for recursion without stack overflows...
except that Rust doesn't have any way of guaranteeing this yet, so you
can't really rely on it.