Implement a Custom Color Picker
The `react-color-picker` library has been unmaintained for quite a
while and it (among other libraries) has been holding us back from
upgrading React properly. I have written a color picker for our Druid
based desktop version before and if we are going to use Dioxus for the
desktop version, I would want to port the Druid version over anyway. And
considering Dioxus is so close to React, I did just that. I ported it to
Dioxus and then let Copilot port it to React, which worked on the very
first try.