Remove `splits.io` Integration (#1027)
As explained [here](https://twos.dev/splitsio.html), splits.io will shut
down at the end of the month. The website is already not working
properly anymore, so it's time to remove support for it.

Changelog: [Splits.io](https://splits.io) is shutting down at the end of
the month. Integration with it has been removed.

[Ben Carlsson's Goodbye to Splits.io](https://twos.dev/splitsio.html)