Allow Popping Out the Timer Into a Separate Window (#1046)
This adds a new button to the sidebar that allows the user to pop out
the entire layout into a separate child window. This makes it take up
less of the screen and capture it more easily in OBS. I'm not sure if
the button should be in the sidebar or not, but it's good enough for an
initial implementation.

Changelog: You can now pop out the timer into a separate window.