Use Document Picture-in-Picture API if Available (#1047)
Chrome supports the [Document Picture-in-Picture
API](https://developer.chrome.com/blog/the-future-of-picture-in-picture/),
which allows us to create a popup window that is always on top, has a
"return to tab" button, remembers its position and size, and has a more
compact title bar. Considering not every browser supports this API, we
automatically fall back to `window.open` if the API is not available.