Fix Bugs Introduced in Refactoring (#1049)
The splits selection view did not correctly cause rerenders among other
smaller bugs in there. I'm not sure how these bugs got introduced, as I
barely made changes to the content of the functions, but these were
completely different. Maybe this was Copilot's fault.

Additionally, when using Tauri, there was a bug where we had two hotkey
implementations at the same time for the pop out windows. One was the
global hotkey implementation and one was the local one we add to the pop
out window. This caused the hotkey to be triggered twice.