Fix Sidebar Not Refreshing in Splits Selection (#1050)
When the splits selection view is open, the sidebar and the view itself
don't refresh properly anymore when in mobile mode. This is because the
React compiler notices that with the way we wrote the code, these
components have no reason to re-render. There is currently no proper
`useMobile` hook. So while I implemented a quick workaround here, by
forcing it to not memoize the individual props properly, we should
refactor the sidebar code at some point to use a proper `useMobile`
hook.