Fix Panic When Cleaning Sum of Best (#1051)
When cleaning the sum of best, the message tries to be as helpful as
possible by showing as much information about the problematic segments
as possible. Part of that is information about the date of the attempt.
However, as it turns out, the attempt may not actually exist. This
shouldn't generally happen, but we can't trust the correctness of the
splits files, so for whatever reason, the attempt may be missing. In
that case, we simply skip that information.